requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-crontab==0.7.1
django-environ==0.9.0
djangorestframework==3.14.0
drf-yasg==1.21.4
et-xmlfile==1.1.0
filelock==3.8.0
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.14.0
google-auth-httplib2==0.1.0
google-cloud-language==2.6.1
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.50.0
httplib2==0.21.0
huggingface-hub==0.10.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.23.3
oauthlib==3.2.1
openai==0.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
pandas-stubs==1.5.0.221010
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.9
psycopg2-binary==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.4
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.9.2
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.3
threadpoolctl==3.1.0
tokenizers==0.13.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.23.1
types-cryptography==3.3.23
types-pytz==2022.4.0.0
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
